refactor: query content through the shared #nuxtseo/content adapter
The @nuxt/content and comark routes were 70 and 49 lines of the same shape
around two manifest formats. nuxtseo-shared normalizes both, so this is one
route file for either provider.

The serverless SQLite hint stays @nuxt/content specific, behind the provider tag
the shim exports rather than behind a separate file.
