+++ src/utils-internal/kit.ts
@@ -1,2 +1,2 @@
 export type { ContentProvider } from 'nuxtseo-shared/kit'
-export { COMARK_CONTENT_MODULE, createNitroPromise, createPagesPromise, getNuxtModuleOptions, isNuxtGenerate, resolveContentProvider, resolveNitroPreset, resolveNuxtContentVersion } from 'nuxtseo-shared/kit'
+export { COMARK_CONTENT_MODULE, createNitroPromise, createPagesPromise, getNuxtModuleOptions, isNuxtGenerate, resolveContentProvider, resolveNitroPreset, resolveNuxtContentVersion, setupContentRuntime } from 'nuxtseo-shared/kit'
